check2FACodeValidity and generate2FASecret are now methods in AuthService, instead of UsersService
